# contracts/openapi/工厂操作系统-v1.1.yaml
openapi: 3.1.0
info:
  title: FactoryOS Platform API
  version: 1.1.1
  description: |
    底座 API v1.1.1 — Core 1.0 + GIP 完整契约（ADR-004/005）。
    Schema 外引：文档/数据结构/*.schema.json
    动词真源：CMV注册表.yaml
    验收：AC-BASE-001 v0.2（**52 P0** + M-03 P1 钩子）
    D1：Graph 版本化路径 · Override/Pack Registry · Blueprint validate · PerceptionOutput
    F1：GET /v1/executions/{execId}/evidence · ExecutionEvidence.schema.json
    F1.5：AgentModule Schema-only（无 HTTP）· DomainEvent 信封 · integration/catalog/
servers:
  - url: http://localhost:8000
    description: FastAPI 开发（Uvicorn）
...
paths:
...
  /v1/rulesets:
    get:
      summary: 列出 RuleSet
      operationId: listRulesets
      tags: [Rule]
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema: { type: string }
        - name: graph_id
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleSet'
    post:
      summary: 创建 draft RuleSet
      operationId: createRuleset
      tags: [Rule]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleSet'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSet'
  /v1/rulesets/{rulesetId}:
    get:
      summary: 获取 RuleSet
      operationId: getRuleset
      tags: [Rule]
      parameters:
        - $ref: '#/components/parameters/rulesetId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSet'
    put:
      summary: 更新 draft RuleSet
      operationId: updateRuleset
      tags: [Rule]
      parameters:
        - $ref: '#/components/parameters/rulesetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleSet'
      responses:
        '200': { description: OK }
        '409': { description: RuleSet frozen }
  /v1/rulesets/{rulesetId}/freeze:
    post:
      summary: 冻结 RuleSet
      operationId: freezeRuleset
      tags: [Rule]
      parameters:
        - $ref: '#/components/parameters/rulesetId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSet'
        '409': { description: Invalid state or graph_version mismatch }
  /v1/rulesets/{rulesetId}/evaluate:
    post:
      summary: 评估规则
      operationId: evaluateRules
      tags: [Rule]
      parameters:
        - $ref: '#/components/parameters/rulesetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleEvaluateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleEvaluateResponse'
...
  /v1/executions/{execId}:
    get:
      summary: 查询 ExecutionRecord
      operationId: getExecution
      tags: [Execution]
      parameters:
        - $ref: '#/components/parameters/execId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionRecord'
...
  /v1/executions/{execId}/evidence:
    get:
      summary: 可重建审计包（Execution + Audit + Rule 快照）
      operationId: getExecutionEvidence
      tags: [Execution]
      parameters:
        - $ref: '#/components/parameters/execId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEvidence'
        '404': { description: Execution not found }
...
  /v1/registry/change-requests:
    get:
      summary: 列出变更请求
      operationId: listRegistryChangeRequests
      tags: [Registry]
      parameters:
        - name: tenant_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigChangeRequest'
    post:
      summary: 提交变更提案（pending）
      operationId: createRegistryChangeRequest
      tags: [Registry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigChangeRequestCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigChangeRequest'
  /v1/registry/change-requests/{requestId}:
    get:
      summary: 变更请求详情
      operationId: getRegistryChangeRequest
      tags: [Registry]
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigChangeRequest'
        '404': { description: Not found }
  /v1/registry/change-requests/{requestId}/approve:
    post:
      summary: 人审批准并应用
      operationId: approveRegistryChangeRequest
      tags: [Registry]
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestDecision'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigChangeRequest'
        '409': { description: Conflict }
  /v1/registry/change-requests/{requestId}/reject:
    post:
      summary: 人审拒绝
      operationId: rejectRegistryChangeRequest
      tags: [Registry]
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestDecision'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigChangeRequest'
components:
  parameters:
...
    execId:
      name: execId
      in: path
      required: true
      schema: { type: string, format: uuid }
    rulesetId:
      name: rulesetId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    RuleSet:
      $ref: '../schemas/规则集.schema.json'
...
    ExecutionRecord:
      $ref: '../schemas/执行记录.schema.json'
    ExecutionEvidence:
      $ref: '../schemas/ExecutionEvidence.schema.json'
...
    RuleEvaluateRequest:
      type: object
      required: [graph_id, graph_version, verb, actor]
      properties:
        graph_id: { type: string }
        graph_version: { type: string }
        verb: { type: string }
        actor:
          type: object
          required: [user_id, role]
          properties:
            user_id: { type: string }
            role: { type: string }
    RuleEvaluateResponse:
      type: object
      required: [effect]
      properties:
        effect: { type: string, enum: [allow, deny] }
        matched_rule_id: { type: string }
...
    ConfigChangeRequestCreate:
      type: object
      required: [kind, proposed_by, proposal_body]
      properties:
        tenant_id: { type: string, nullable: true }
        kind: { type: string, enum: [pack_upsert, system_relation_upsert] }
        proposed_by: { type: string }
        proposal_body: { type: object }
        ai_model_id: { type: string, nullable: true }
    ConfigChangeRequest:
      type: object
      required: [request_id, kind, status, proposed_by, proposal_body]
      properties:
        request_id: { type: string }
        tenant_id: { type: string, nullable: true }
        kind: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        proposed_by: { type: string }
        proposal_body: { type: object }
        ai_model_id: { type: string, nullable: true }
        created_at: { type: string }
    ChangeRequestDecision:
      type: object
      required: [actor_id]
      properties:
        actor_id: { type: string }
        reason: { type: string, nullable: true }
...